+++ src/Schema.php
@@ -35,6 +35,7 @@ use LaravelJsonApi\Eloquent\Fields\Relations\ToMany;
 use LaravelJsonApi\Eloquent\Fields\Relations\ToOne;
 use LaravelJsonApi\Eloquent\Parsers\StandardParser;
 use LogicException;
+use function sprintf;
 
 abstract class Schema extends BaseSchema implements CountableSchema
 {
@@ -52,6 +53,13 @@ abstract class Schema extends BaseSchema implements CountableSchema
      * @var array|null
      */
     protected ?array $defaultPagination = null;
+
+    /**
+     * The default sort order for this resource.
+     *
+     * @var string|string[]|null
+     */
+    protected $defaultSort = null;
 
     /**
      * The cached parser instance.
@@ -253,6 +261,20 @@ abstract class Schema extends BaseSchema implements CountableSchema
         }
 
         return $this->defaultEagerLoadPaths = array_values(array_unique($paths));
+    }
+
+    /**
+     * Get the default sort order for this resource.
+     *
+     * The default sort order is used if the client supplies no sort parameters.
+     * Returning `null` from this method indicates that no default sort order exists
+     * and resource should be returned in the order they are retrieved from the database.
+     *
+     * @return string|string[]|null
+     */
+    public function defaultSort()
+    {
+        return $this->defaultSort;
     }
 
     /**