+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- Schemas now support additional sort field classes, that define how to sort models using sort fields that are not
+  attributes. Sort field classes must implement the new `SortField` contract. Two initial sort classes are available:
+  `SortColumn` and `SortCount`.
+- Default sort order for resources can now be defined on the Eloquent schema using the `$defaultSort` property.
+
 ## [1.0.0-beta.3] - 2021-04-22
 
 ### Added